refactor(pack): refactor xdsl statistics
