+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/pack-xdsl-access.html
@@ -871,7 +871,7 @@
                         data-translate="xdsl_statistics_title"
                     ></h3>
                     <div
-                        data-ng-include="'app/telecom/pack/xdsl/access/statistics/pack-xdsl-statistics.html'"
+                        data-ng-include="'app/telecom/pack/xdsl/access/statistics/statistics.html'"
                     ></div>
                 </div>
             </div>
