+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/statistics/statistics.html
